rename to agent AC, add position class
