+++ soccer/CMakeLists.txt
@@ -85,7 +85,7 @@ add_executable(sim_radio_node)
 add_executable(network_radio_node)
 add_executable(manual_control_node)
 add_executable(global_param_server_node)
-add_executable(robot_action_client_node)
+add_executable(agent_action_client_node)
 
 if(BUILD_TESTING)
   add_executable(test-soccer)
@@ -184,8 +184,8 @@ set(MANUAL_CONTROL_NODE_DEPS_LIBRARIES robocup)
 set(GLOBAL_PARAM_SERVER_NODE_DEPS_SYSTEM_LIBRARIES)
 set(GLOBAL_PARAM_SERVER_NODE_DEPS_LIBRARIES robocup)
 
-set(ROBOT_ACTION_CLIENT_NODE_DEPS_SYSTEM_LIBRARIES)
-set(ROBOT_ACTION_CLIENT_NODE_DEPS_LIBRARIES robocup)
+set(AGENT_ACTION_CLIENT_NODE_DEPS_SYSTEM_LIBRARIES)
+set(AGENT_ACTION_CLIENT_NODE_DEPS_LIBRARIES robocup)
 
 # ======================================================================
 # Include and Linking
@@ -245,9 +245,9 @@ target_link_libraries(manual_control_node PRIVATE ${MANUAL_CONTROL_NODE_DEPS_SYS
 target_link_libraries(global_param_server_node PRIVATE ${GLOBAL_PARAM_SERVER_NODE_DEPS_SYSTEM_LIBRARIES}
         ${GLOBAL_PARAM_SERVER_NODE_DEPS_LIBRARIES})
 
-# -- robot_action_client_node --
-target_link_libraries(robot_action_client_node PRIVATE ${ROBOT_ACTION_CLIENT_NODE_DEPS_SYSTEM_LIBRARIES}
-    ${ROBOT_ACTION_CLIENT_NODE_DEPS_LIBRARIES})
+# -- agent_action_client_node --
+target_link_libraries(agent_action_client_node PRIVATE ${AGENT_ACTION_CLIENT_NODE_DEPS_SYSTEM_LIBRARIES}
+    ${AGENT_ACTION_CLIENT_NODE_DEPS_LIBRARIES})
 
 # ======================================================================
 # Testing
@@ -292,7 +292,7 @@ install(
         network_radio_node
         manual_control_node
         global_param_server_node
-        robot_action_client_node
+        agent_action_client_node
         DESTINATION lib/${CMAKE_PROJECT_NAME})
 
 install(
