+++ soccer/src/soccer/CMakeLists.txt
@@ -68,7 +68,9 @@ set(ROBOCUP_LIB_SRC
     ros2_temp/referee_sub.cpp
     ros2_temp/autonomy_interface.cpp
     ros2_temp/debug_draw_interface.cpp
-    robot_action_client/robot_action_client.cpp
+    agent_action_client/agent_action_client.cpp
+    agent_action_client/position/position.cpp
+    agent_action_client/position/goalie.cpp
         )
 
 set(SOCCER_TEST_SRC
@@ -113,7 +115,7 @@ set(MANUAL_CONTROL_NODE_SRC joystick/manual_control_node_main.cpp)
 
 set(GLOBAL_PARAMETER_SERVER_NODE_SRC global_param_server.cpp)
 
-set(ROBOT_ACTION_CLIENT_NODE_SRC robot_action_client/robot_action_client_main.cpp)
+set(AGENT_ACTION_CLIENT_NODE_SRC agent_action_client/agent_action_client_main.cpp)
 
 # ======================================================================
 # Add sources
@@ -155,8 +157,8 @@ target_sources(manual_control_node PRIVATE ${MANUAL_CONTROL_NODE_SRC})
 # ---- global_param_server_node ----
 target_sources(global_param_server_node PRIVATE ${GLOBAL_PARAMETER_SERVER_NODE_SRC})
 
-# ---- robot_action_client_node ----
-target_sources(robot_action_client_node PRIVATE ${ROBOT_ACTION_CLIENT_NODE_SRC})
+# ---- agent_action_client_node ----
+target_sources(agent_action_client_node PRIVATE ${AGENT_ACTION_CLIENT_NODE_SRC})
 
 # Set the directory so AUTOUIC know where to find .ui files
 set_target_properties(log_viewer PROPERTIES AUTOUIC_SEARCH_PATHS ${UI_PATH})
